Object.m (errno): Replaced by errno.h include.
2008-11-21  Kai Tietz  <kai.tietz@onevision.com>

        *  Object.m (errno): Replaced by errno.h include.
        (compare): Cast self to id to prevent warning on comparison.
        * objc/objc.h (BOOL): Prevent redeclaration of BOOL, if it is
        already there.
        * sendmsg.c (__objc_print_dtable_stats): Remove type warnings.
        * thr-win32.c (__objc_thread_detach): Remove type warning.
        (__objc_thread_id): Likewise.
        * thr.c (__objc_thread_detach_functiont): Add __builtin_trap ()
        for noreturn.

From-SVN: r142087

+++ libobjc/sendmsg.c
@@ -687,14 +687,14 @@ __objc_print_dtable_stats ()
 #endif
 
   printf ("arrays: %d = %ld bytes\n", narrays, 
-	  (long) (narrays * sizeof (struct sarray)));
+	  (long) ((size_t) narrays * sizeof (struct sarray)));
   total += narrays * sizeof (struct sarray);
   printf ("buckets: %d = %ld bytes\n", nbuckets, 
-	  (long) (nbuckets * sizeof (struct sbucket)));
+	  (long) ((size_t) nbuckets * sizeof (struct sbucket)));
   total += nbuckets * sizeof (struct sbucket);
 
   printf ("idxtables: %d = %ld bytes\n",
-	  idxsize, (long) (idxsize * sizeof (void *)));
+	  idxsize, (long) ((size_t) idxsize * sizeof (void *)));
   total += idxsize * sizeof (void *);
   printf ("-----------------------------------\n");
   printf ("total: %d bytes\n", total);
